+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Added `"Optimal dispatch metadata"` and `"Optimal switching metadata"` to the output schema
 - Added arguments `--opt-switch-algorithm`, `--opt-switch-solver`, `--opt-disp-algorithm`
 - Changed `optimize_dispatch!` function to not overwrite base network data by default
 - Added `"mip_gap"` to the results dictionary, if it exists